// bulgaria-travel-gems/src/app/types/province.enum.ts
export enum Province {
  Blagoevgrad = 'Blagoevgrad',
  Burgas = 'Burgas',
  Dobrich = 'Dobrich',
  Gabrovo = 'Gabrovo',
  Haskovo = 'Haskovo',
  Kardzhali = 'Kardzhali',
  Kyustendil = 'Kyustendil',
  Lovech = 'Lovech',
  Montana = 'Montana',
  Pazardzhik = 'Pazardzhik',
  Pernik = 'Pernik',
  Pleven = 'Pleven',
  Plovdiv = 'Plovdiv',
  Razgrad = 'Razgrad',
  Ruse = 'Ruse',
  Shumen = 'Shumen',
  Silistra = 'Silistra',
  Sliven = 'Sliven',
  Smolyan = 'Smolyan',
  SofiaCity = 'Sofia City',
  SofiaProvince = 'Sofia Province',
  StaraZagora = 'Stara Zagora',
  Targovishte = 'Targovishte',
  Varna = 'Varna',
  VelikoTarnovo = 'Veliko Tarnovo',
  Vidin = 'Vidin',
  Vratsa = 'Vratsa',
  Yambol = 'Yambol',
}